Add columns that may be out of order. Fixes issue #8.
 * pyrseas/dbobject/column.py (ColumnDict.diff_map): Don't attempt to
   drop columns that have already been dropped in the catalogs (caused
   innocuous semicolons on output).
 * pyrseas/dbobject/table.py (Table.diff_map): Examine existing
   columns and generate ADD COLUMN statements if not present in the
   input map.

+++ pyrseas/dbobject/column.py
@@ -183,7 +183,8 @@ class ColumnDict(DbObjectDict):
         for (sch, tbl) in incols.keys():
             if (sch, tbl) in self.keys():
                 for col in self[(sch, tbl)]:
-                    if col.name not in [c.name for c in incols[(sch, tbl)]]:
+                    if col.name not in [c.name for c in incols[(sch, tbl)]] \
+                            and not hasattr(col, 'dropped'):
                         stmts.append(col.drop())
 
         return stmts
